General updates to sample_size_baseline_and_cohort shiny app
I rewrote most of the app trying to keep the interface similar to the
original version. Plots modified, added tables, interface is now fully
reactive, added one extra tab with power plot as function of incidence
reduction %,. Most of the underlying code has been rewritten and when
possible functions split in subfunctions for easier handling and
debugging. Also reduced the number of points needed to draw the power
curve, so the app can be faster.